+++ src/AllGather.cpp
@@ -146,9 +146,9 @@ int main(int argc, char *argv[]) {
 
     int BUFFER_SIZE = (size_count * multiplier_type);
       int MSG_BUFFER_SIZE = BUFFER_SIZE / size;
-      int msg_count = MSG_BUFFER_SIZE / sizeof(float);
-      float *send_buffer = (float*) malloc(MSG_BUFFER_SIZE); 
-      float *recv_buffer = (float*) malloc(BUFFER_SIZE);
+      int msg_count = MSG_BUFFER_SIZE / sizeof(int);
+      int *send_buffer = (int*) malloc(MSG_BUFFER_SIZE);
+      int *recv_buffer = (int*) malloc(BUFFER_SIZE);
     if (send_buffer == NULL || recv_buffer == NULL) {
         fprintf(stderr, "Memory allocation failed!\n");
         MPI_Abort(MPI_COMM_WORLD, 1);
@@ -156,7 +156,7 @@ int main(int argc, char *argv[]) {
     }
 
     for (int i = 0; i < msg_count; i++) {
-        send_buffer[i] = (float) (rand()*rank % 10);
+        send_buffer[i] = (int) (rand()*rank % 10);
     }
 
     double* samples = (double*) malloc(sizeof(double) * BENCHMARK_ITERATIONS);
@@ -166,7 +166,7 @@ int main(int argc, char *argv[]) {
 
         double start_time, end_time;
         start_time = MPI_Wtime();
-        MPI_Allgather(send_buffer, msg_count, MPI_FLOAT, recv_buffer, msg_count, MPI_FLOAT, MPI_COMM_WORLD);
+        MPI_Allgather(send_buffer, msg_count, MPI_INT, recv_buffer, msg_count, MPI_INT, MPI_COMM_WORLD);
         end_time = MPI_Wtime();
 
         if(i>WARM_UP) {
